+++ src/rcpp_slice.cpp
@@ -11,19 +11,19 @@ using namespace Rcpp;
 // [[Rcpp::plugins(cpp11)]]
 
 struct Region {
-    unsigned long start, end, value;
+    long start, end, value;
     Region (): start(0), end(0), value(0) {}
-    Region (unsigned long s, unsigned long e, unsigned long v):
+    Region (long s, long e, long v):
         start(s), end(e), value(v) {}
     bool empty () {
         return (start == 0);
     }
-    unsigned long getSize () {
+    long getSize () {
         return end - start + 1;
     }
-    unsigned long getDifference (Region &r) {
+    long getDifference (Region &r) {
         if ((empty()) || (r.empty()))
-            return std::numeric_limits<unsigned long>::max();
+            return std::numeric_limits<long>::max();
         return ((start >= r.start)? start - r.start: r.start - start) +
                ((end   >= r.end  )? end   - r.end  : r.end   - end);
     }
@@ -34,9 +34,9 @@ struct Region {
     }
 };
 
-double computeMedian(std::vector < std::pair < double, unsigned int > > &table){
+double computeMedian(std::vector < std::pair < double, int > > &table){
     sort(table.begin(), table.end());
-    size_t i, s = 0, size = 0;
+    int i, s = 0, size = 0;
     for (auto &i: table) {
             size += i.second;
     }
@@ -59,19 +59,19 @@ void rcpp_normalization(ListOf < ListOf < IntegerVector > > &lengths,
                         IntegerVector &chromosomeSizes,
                         IntegerVector &librarySizes) {
     GenomeIterator iterator (lengths, values, chromosomeSizes);
-    std::vector < std::pair < std::valarray < double >, unsigned int > >
+    std::vector < std::pair < std::valarray < double >, int > >
         normValues;
-    std::vector < std::pair < double, unsigned int > > normValuesPerSample;
-    std::vector < std::pair < double, unsigned int > > avgValues;
-    std::vector < std::pair < double, unsigned int > > avgNormValues;
-    size_t nSamples     = lengths[0].size();
-    size_t nChromosomes = chromosomeSizes.size();
+    std::vector < std::pair < double, int > > normValuesPerSample;
+    std::vector < std::pair < double, int > > avgValues;
+    std::vector < std::pair < double, int > > avgNormValues;
+    int nSamples     = lengths[0].size();
+    int nChromosomes = chromosomeSizes.size();
     std::valarray < double > theseValues (nSamples);
     std::valarray < double > factors (nSamples);
     std::valarray < double > sums (nSamples);
     std::valarray < double > normalizedSums (nSamples);
     std::valarray < double > librarySizesArray (nSamples);
-    for (unsigned int i = 0; i < nSamples; ++i) {
+    for (int i = 0; i < nSamples; ++i) {
         librarySizesArray[i] = librarySizes[i];
     }
     for (; ! iterator.isOver(); iterator.getNext()) {
@@ -81,7 +81,7 @@ void rcpp_normalization(ListOf < ListOf < IntegerVector > > &lengths,
             normValues.emplace_back(theseValues, iterator.getStep());
         }
     }
-    for (size_t sample = 0; sample < nSamples; ++sample) {
+    for (int sample = 0; sample < nSamples; ++sample) {
         normValuesPerSample.clear();
         normValuesPerSample.reserve(normValues.size());
         for (auto &normValue: normValues) {
@@ -103,11 +103,11 @@ void rcpp_normalization(ListOf < ListOf < IntegerVector > > &lengths,
                      std::begin(normalizedSums) + nSamples / 2,
                      std::end(normalizedSums));
     double fs     = sums[nSamples / 2] / normalizedSums[nSamples / 2];
-    for (size_t chromosomeId = 0; chromosomeId < nChromosomes; ++chromosomeId) {
-        for (size_t sampleId = 0; sampleId < nSamples; ++sampleId) {
-            for (size_t i = 0; i < values[chromosomeId][sampleId].size(); ++i) {
+    for (int chromosomeId = 0; chromosomeId < nChromosomes; ++chromosomeId) {
+        for (int sampleId = 0; sampleId < nSamples; ++sampleId) {
+            for (int i = 0; i < values[chromosomeId][sampleId].size(); ++i) {
                 values[chromosomeId][sampleId][i] =
-                    static_cast<unsigned int>(round(static_cast<double>(
+                    static_cast<int>(round(static_cast<double>(
                             values[chromosomeId][sampleId][i]) * fs));
             }
         }
@@ -130,46 +130,46 @@ List rcpp_slice(ListOf < ListOf < IntegerVector > > &lengths,
                    IntegerVector &chromosomeSizes, int minDepth,
                    int minSize, int maxSize, int minDifference) {
     static const short nConditions = 2;
-    std::vector < unsigned int > starts, ends;
+    std::vector < int > starts, ends;
     std::vector < std::string > chromosomes;
-    size_t nChromosomes = chromosomeSizes.size();
+    int nChromosomes = chromosomeSizes.size();
     std::vector < Region > selectedRegions[nConditions];
     std::string chromosome;
-    for (size_t chrId = 0; chrId < nChromosomes; ++chrId) {
+    for (int chrId = 0; chrId < nChromosomes; ++chrId) {
         chromosome = as < std::string >(as < CharacterVector> (
             chromosomeSizes.names())[chrId]);
         for (int condition = 0; condition < nConditions; ++condition) {
             std::vector < Region > theseRegions;
-            size_t        nValues = lengths[condition][chrId].size();
-            unsigned long start   = 1;
-            int           rLast   = 0;
+            int  nValues = lengths[condition][chrId].size();
+            long start   = 1;
+            int  rLast   = 0;
             selectedRegions[condition].clear();
-            for (size_t index = 0; index < nValues; ++index) {
-                unsigned long length    = lengths[condition][chrId][index];
-                unsigned long value     = values[condition][chrId][index];
-                unsigned long end       = start + length - 1;
-								int           nextRLast = theseRegions.size();
+            for (int index = 0; index < nValues; ++index) {
+                long length    = lengths[condition][chrId][index];
+                long value     = values[condition][chrId][index];
+                long end       = start + length - 1;
+                int  nextRLast = theseRegions.size();
                 for (int rit = theseRegions.size()-1; rit >= rLast; --rit) {
-										Region &region = theseRegions[rit];
+                    Region &region = theseRegions[rit];
                     if ((region.end == start - 1) && (region.value <= value)) {
                         region.end = end;
-												nextRLast = rit;
+                        nextRLast = rit;
                     }
                 }
-								rLast = nextRLast;
+                rLast = nextRLast;
                 if (value >= minDepth) {
                     theseRegions.emplace_back(start, end, value);
                 }
                 start = end + 1;
             }
             for (Region &region: theseRegions) {
-                unsigned long size = region.getSize();
+                long size = region.getSize();
                 if ((minSize <= size) && (size <= maxSize)) {
-                    unsigned long difference = minDifference + 1;
+                    long difference = minDifference + 1;
                     for (auto rit = selectedRegions[condition].rbegin();
                          rit != selectedRegions[condition].rend(); ++rit) {
-                        unsigned long d = region.getDifference(*rit);
-                        difference = std::min<unsigned long>(difference, d);
+                        long d = region.getDifference(*rit);
+                        difference = std::min<long>(difference, d);
                         if (d > 10 * minDifference) {
                             break;
                         }
@@ -187,14 +187,14 @@ List rcpp_slice(ListOf < ListOf < IntegerVector > > &lengths,
             {selectedRegions[0].begin(), selectedRegions[1].begin()};
         while ((its[0] != selectedRegions[0].end()) &&
                (its[1] != selectedRegions[1].end())) {
-            unsigned int condition = (*its[0] < *its[1])? 0: 1;
+            int condition = (*its[0] < *its[1])? 0: 1;
             mergedRegions.push_back(*its[condition]);
             ++its[condition];
         }
         for (int condition = 0; condition < nConditions; ++condition) {
             while (its[condition] != selectedRegions[condition].end()) {
                 mergedRegions.push_back(*its[condition]);
-								++its[condition];
+                ++its[condition];
             }
         }
         Region previousRegion;
